requirements.txt
Flask==2.3.3
pandas==2.0.3
matplotlib==3.7.2
python-dotenv==1.0.1
streamlit==1.24.0